requirements.txt
anitya_schema==2.3.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
Automat==25.4.16
blinker==1.9.0
boto3==1.40.74
botocore==1.40.74
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
constantly==23.10.4
crochet==2.1.1
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
fastapi==0.121.2
fedora-messaging-the-new-hotness-schema==1.2.1
fedora_messaging==3.9.0
greenlet==3.2.4
gssapi==1.10.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jiter==0.12.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
koji==1.35.3
openai==2.8.0
paho-mqtt==2.1.0
pika==1.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
PyMySQL==1.1.2
pyOpenSSL==25.3.0
python-bugzilla==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==5.3.1
referencing==0.37.0
requests==2.32.5
requests-gssapi==1.4.0
rpds-py==0.29.0
s3transfer==0.14.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
the-new-hotness==1.3.0
tomli==2.3.0
tqdm==4.67.1
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websocket-client==1.9.0
wrapt==2.0.1
zope.interface==8.1.1
